Move to NavigationBar.tsx, simplify min max zoom
